backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
httpx==0.26.0
pydantic-settings==2.1.0
asyncpg==0.30.0
python-dotenv==1.0.1

# Auth (FastAPI Users 13.x)
fastapi-users[sqlalchemy]==13.0.0
pwdlib[argon2]==0.2.1
python-jose[cryptography]==3.3.0
redis==5.0.0
sqlalchemy[asyncio]==2.0.0
aiosqlite==0.20.0
email-validator>=2.1.1

# Document parsing (replaces PyMuPDF)
docling==2.77.0
pypdf==5.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
